Conflicts:
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/dao/ChartDetailDao.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/AccountLog.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/User.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/UserDeviceDetails.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/payment/PSMSPaymentDetails.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/domain/payment/PaymentDetails.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/repository/UserGroupRepository.java
	persistence/src/main/java/mobi/nowtechnologies/server/persistence/repository/UserRepository.java
	persistence/src/main/resources/META-INF/jpql/chartDetail.xml
	persistence/src/main/resources/sql/release5.0-SNAPSHOT.sql
	persistence/src/test/java/mobi/nowtechnologies/server/persistence/domain/UserTest.java
	persistence/src/test/java/mobi/nowtechnologies/server/persistence/repository/UserRepositoryIT.java
	service/src/main/java/mobi/nowtechnologies/server/service/UserService.java
	service/src/main/java/mobi/nowtechnologies/server/service/impl/UserNotificationServiceImpl.java
	service/src/test/java/mobi/nowtechnologies/server/service/PaymentDetailsServiceTest.java
	service/src/test/java/mobi/nowtechnologies/server/service/UserServiceTest.java
	service/src/test/java/mobi/nowtechnologies/server/service/impl/UserNotificationServiceImplTest.java
	transport/src/main/java/mobi/nowtechnologies/server/job/ForkO2UsersForUpdateJob.java
	transport/src/main/webapp/WEB-INF/conf/log4j.xml

+++ persistence/src/test/java/mobi/nowtechnologies/server/persistence/repository/UserRepositoryIT.java
@@ -41,7 +41,7 @@ public class UserRepositoryIT {
 	private static final int HOUR_SECONDS = 60 * 60;
 	private static final int DAY_SECONDS = 24 * HOUR_SECONDS;
 	private static final int TWO_DAY_SECONDS = 2 * DAY_SECONDS;
-	private static final byte o2CommunityId = 7;
+	private static final int o2CommunityId = 7;
 
 	@Resource(name = "userRepository")
 	private UserRepository userRepository;
